add in test for func
