+++ beacon-chain/core/helpers/committee.go
@@ -363,3 +363,42 @@ func AttestationParticipants(
 	}
 	return participants, nil
 }
+
+// VerifyBitfield validates a bitfield with a given committee size.
+//
+// Spec pseudocode:
+//
+// def verify_bitfield(bitfield: bytes, committee_size: int) -> bool:
+// """
+// Verify ``bitfield`` against the ``committee_size``.
+// """
+// if len(bitfield) != (committee_size + 7) // 8:
+// return False
+//
+// # Check `bitfield` is padded with zero bits only
+// for i in range(committee_size, len(bitfield) * 8):
+// if get_bitfield_bit(bitfield, i) == 0b1:
+// return False
+//
+// return True
+func VerifyBitfield(bitfield []byte, committee_size int) (bool, error) {
+	if len(bitfield) != mathutil.CeilDiv8(committee_size) {
+		return false, fmt.Errorf(
+			"wanted participants bitfield length %d, got: %d",
+			mathutil.CeilDiv8(committee_size),
+			len(bitfield))
+	}
+
+	for i := committee_size; i < len(bitfield)*8; i++ {
+		bitSet, err := bitutil.CheckBit(bitfield, i)
+		if err != nil {
+			return false, fmt.Errorf("unable to check bit in bitfield %v", err)
+		}
+
+		if bitSet {
+			return false, nil
+		}
+	}
+
+	return true, nil
+}
