+++ beacon-chain/core/helpers/signature.go
@@ -1,12 +1,9 @@
 package helpers
 
 import (
-	"fmt"
 	"math"
 
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
-	"github.com/prysmaticlabs/prysm/shared/bitutil"
-	"github.com/prysmaticlabs/prysm/shared/mathutil"
 )
 
 // ForkVersion returns the fork version of the given epoch number.
@@ -42,43 +39,4 @@ func ForkVersion(fork *pb.Fork, epoch uint64) uint64 {
 func DomainVersion(fork *pb.Fork, epoch uint64, domainType uint64) uint64 {
 	offset := uint64(math.Pow(2, 32))
 	return ForkVersion(fork, epoch)*offset + domainType
-}
-
-// VerifyBitfield validates a bitfield with a given committee size.
-//
-// Spec pseudocode:
-//
-// def verify_bitfield(bitfield: bytes, committee_size: int) -> bool:
-// """
-// Verify ``bitfield`` against the ``committee_size``.
-// """
-// if len(bitfield) != (committee_size + 7) // 8:
-// return False
-//
-// # Check `bitfield` is padded with zero bits only
-// for i in range(committee_size, len(bitfield) * 8):
-// if get_bitfield_bit(bitfield, i) == 0b1:
-// return False
-//
-// return True
-func VerifyBitfield(bitfield []byte, committee_size int) (bool, error) {
-	if len(bitfield) != mathutil.CeilDiv8(committee_size) {
-		return false, fmt.Errorf(
-			"wanted participants bitfield length %d, got: %d",
-			mathutil.CeilDiv8(committee_size),
-			len(bitfield))
-	}
-
-	for i := committee_size; i < len(bitfield); i++ {
-		bitSet, err := bitutil.CheckBit(bitfield, i)
-		if err != nil {
-			return false, fmt.Errorf("unable to check bit in bitfield %v", err)
-		}
-
-		if !bitSet {
-			return false, nil
-		}
-	}
-
-	return true, nil
 }
